Added gamblingScript
